Move chart to a component
